refactor: moe SignerConf into configuration crate
