+++ configuration/src/agent/signer.rs
@@ -1,0 +1,29 @@
+use nomad_types::HexString;
+
+/// Ethereum signer types
+#[derive(Debug, Clone, PartialEq, serde::Deserialize)]
+#[serde(tag = "type", rename_all = "camelCase")]
+pub enum SignerConf {
+    /// A local hex key
+    HexKey {
+        /// Hex string of private key, without 0x prefix
+        key: HexString<64>,
+    },
+    /// An AWS signer. Note that AWS credentials must be inserted into the env
+    /// separately.
+    Aws {
+        /// The UUID identifying the AWS KMS Key
+        id: String, // change to no _ so we can set by env
+        /// The AWS region
+        region: String,
+    },
+    #[serde(other)]
+    /// Assume node will sign on RPC calls
+    Node,
+}
+
+impl Default for SignerConf {
+    fn default() -> Self {
+        Self::Node
+    }
+}
