+++ nomad-types/Cargo.toml
@@ -13,3 +13,5 @@ thiserror = "*"
 tracing-subscriber = "0.2.15"
 prometheus = "0.12"
 affix = "0.1.2"
+color-eyre = "0.5.0"
+hex = "0.4.3"
