+++ nomad-types/src/utils.rs
@@ -1,0 +1,12 @@
+/// Strips the '0x' prefix off of hex string so it can be deserialized.
+///
+/// # Arguments
+///
+/// * `s` - The hex str
+pub fn strip_0x_prefix(s: &str) -> &str {
+    if s.len() < 2 || &s[..2] != "0x" {
+        s
+    } else {
+        &s[2..]
+    }
+}
